requirements.txt
autopep8==1.5.7
billiard==3.6.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
docutils==0.18.1
Flask==2.0.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
kombu==5.1.0
lockfile==0.12.2
MarkupSafe==2.0.0
marshmallow==3.12.1
prompt-toolkit==3.0.19
psycopg2==2.8.6
pycodestyle==2.7.0
python-daemon==2.3.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
urllib3==1.26.4
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.0
